Update target to ES2020 and replace Long with bigint (#912)
This commit Streamlines data type handling by replacing Long with native.
It also updates target to ECMAScript 2020.

Despite the introduction of bigint, but Long still remains in Counter. We
need to remove Long.

+++ packages/sdk/test/integration/client_test.ts
@@ -522,12 +522,12 @@ describe.sequential('Client', function () {
 
     let checkpoint = d1.getCheckpoint();
     assert.equal(checkpoint.getClientSeq(), 1);
-    assert.equal(checkpoint.getServerSeq().toInt(), 1);
+    assert.equal(checkpoint.getServerSeq(), 1n);
 
     await c1.sync();
     checkpoint = d1.getCheckpoint();
     assert.equal(checkpoint.getClientSeq(), 2);
-    assert.equal(checkpoint.getServerSeq().toInt(), 2);
+    assert.equal(checkpoint.getServerSeq(), 2n);
 
     // 03. cli update the document with increasing the counter(0 -> 1)
     //     and sync with push-only mode: CP(2, 2) -> CP(3, 2)
@@ -544,7 +544,7 @@ describe.sequential('Client', function () {
     await eventCollector.waitFor(DocumentSyncStatus.Synced);
     checkpoint = d1.getCheckpoint();
     assert.equal(checkpoint.getClientSeq(), 3);
-    assert.equal(checkpoint.getServerSeq().toInt(), 2);
+    assert.equal(checkpoint.getServerSeq(), 2n);
     await c1.changeSyncMode(d1, SyncMode.Manual);
 
     // 04. cli update the document with increasing the counter(1 -> 2)
@@ -561,7 +561,7 @@ describe.sequential('Client', function () {
     await c1.sync();
     checkpoint = d1.getCheckpoint();
     assert.equal(checkpoint.getClientSeq(), 4);
-    assert.equal(checkpoint.getServerSeq().toInt(), 4);
+    assert.equal(checkpoint.getServerSeq(), 4n);
     assert.equal(d1.getRoot().counter.getValue(), 2);
 
     unsub();
